Move timed storage to it's own files.
