+++ ui/src/lib/useTimedStorage/index.ts
@@ -1,24 +1,8 @@
-import { StorageLike, StorageSerializers, UseStorageOptions, useStorage } from '@vueuse/core';
+import { StorageSerializers, useStorage } from '@vueuse/core';
 import { differenceInMilliseconds, parseISO } from 'date-fns';
 import { WritableComputedRef, computed } from 'vue';
-
-export interface UseTimedStorageOptions<T> {
-  /** The value to use as the storage key. Always prefixed with `[nh]` */
-  keyName: string;
-  /** The optional value to use as the initial value. @default null */
-  initialValue?: T | null;
-  /** The time this storage value is allowed to be accessed. @default '5m' */
-  liveTime?: number;
-  /** The base storage interface to use for this hook. @default SessionStorage */
-  storage?: StorageLike;
-  /** The optional extension value to configure the useStorage native behavior. */
-  storageOptions?: UseStorageOptions<T>;
-}
-
-interface TimedStorageVal<T> {
-  storedAt: Date | string;
-  value: T;
-}
+import type { TimedStorageVal } from './timedStorageVal';
+import type { UseTimedStorageOptions } from './useTimedStorageOptions';
 
 export function useTimedStorage<T>(keyName: string): WritableComputedRef<T | null>;
 export function useTimedStorage<T>(
@@ -78,3 +62,6 @@ export function useTimedStorage<T>(
     set: (nv) => (storedVal.value = nv === null ? null : { storedAt: new Date(), value: nv }),
   });
 }
+
+export type { TimedStorageVal } from './timedStorageVal';
+export type { UseTimedStorageOptions } from './useTimedStorageOptions';
